fix test that interceptors are dependent instances of intercepted beans
The test previously used a `@PreDestroy` callback declared on the interceptor
to prove that the interceptor is destroyed when the intercepted bean is
destroyed, which is incorrect. Interceptors cannot declare lifecycle callbacks
for themselves, the `@PreDestroy` callback actually interposed on the destruction
of the intercepted bean.

This commit uses a new dependent-scoped bean that is injected into the interceptor,
which makes it a dependent instance of the interceptor. The new bean has
a `@PreDestroy` callback for itself, which means we can verify that the new bean
is destroyed when the intercepted bean is destroyed. Such chain of dependent
instance destruction proves that the interceptor is indeed a dependent instance
of the intercepted bean.